remove duplicated builtin folder& files
