+++ src/builtin/builtin.c
@@ -6,22 +6,23 @@
 /*   By: bewong <bewong@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/21 15:14:26 by bewong        #+#    #+#                 */
-/*   Updated: 2025/01/22 21:25:03 by bewong        ########   odam.nl         */
+/*   Updated: 2025/01/23 18:49:25 by bewong        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "executor.h"
-#include "minishell.h"
 
-static int	ft_unset(t_exec *cmd)
+static int	ft_unset(t_ast_node *node, t_env **env)
 {
 	t_env	*tmp;
 	int		i;
 
+	if (!node || !env || !*env)
+		return (EXIT_FAILURE);
 	i = 1;
-	while (i < cmd->argc)
+	while (i < node->args_count)
 	{
-		tmp = get_env(*(cmd)->env, cmd->argv[i]);
+		tmp = get_env(*env, node->args[i]);
 		if (tmp)
 			tmp->hide = true;
 		i++;
@@ -29,16 +30,18 @@ static int	ft_unset(t_exec *cmd)
 	return (EXIT_SUCCESS);
 }
 
-static int	builtin_env(t_exec *cmd)
+static int	builtin_env(t_ast_node *node, t_env **env)
 {
 	t_env	*head;
 
-	if (cmd->argc > 1)
+	if (!node || !env || !*env)
+		return (EXIT_FAILURE);
+	if (node->args_count > 1)
 	{
 		ft_putendl_fd("Too many args", STDERR_FILENO);
 		return (EXIT_FAILURE);
 	}
-	head = *(cmd->env);
+	head = *env;
 	while (head)
 	{
 		if (head->hide == false && (head->scope & (BOTH | ENVE)))
@@ -51,66 +54,36 @@ static int	builtin_env(t_exec *cmd)
 /*
 	If the getcwd function fails, it return 1, otherwise returns 1.
 */
-static int	builtin_pwd(t_exec *cmd)
+static int	builtin_pwd(t_ast_node *node, t_env **env)
 {
 	char	cwd[PATH_MAX];
 
-	if (cmd->argc != 1)
+	if (node->args_count != 1)
 		return (ft_putendl_fd(MANY_ARGS_ERROR, STDERR_FILENO), 1);
 	getcwd(cwd, PATH_MAX);
 	printf("%s\n", cwd);
 	return (EXIT_SUCCESS);
 }
-/*
-	echo with -n do not output the trailing newline
-	echo with no args, output a newline
-	echo with multiple -n, skip those -n
-	echo with invalid -n, e.g. -nn, will directly output -nn
-*/
-static int	builtin_echo(t_exec *cmd)
-{
-	int	i;
-	int	no_newline;
 
-	i = 1;
-	no_newline = -1;
-	if (cmd->argv[i] && ft_strncmp(cmd->argv[++i], "-n", 3) == 0)
-		no_newline = 1;
-	while (cmd->argv[i])
-	{
-		if (ft_strcmp(cmd->argv[i], "$?") == 0) //$? case included
-		{
-			printf("%d", get_exit_status());
-			i++;
-			continue;
-		}
-		printf("%s", cmd->argv[i]);
-		if (cmd->argv[i + 1])
-			printf(" ");
-		i++;
-	}
-	if (!no_newline)
-		printf("\n");
-	return (EXIT_SUCCESS);
-}
 
-int	execute_builtin(t_exec *cmd)
+int	execute_builtin(t_ast_node *node, t_env **env)
 {
-	if (!cmd || !cmd->argv || !cmd->argv[0])
-		return (1);
-	if (ft_strcmp(cmd->argv[0], "cd") == 0)
-		return (builtin_cd(cmd));
-	if (ft_strcmp(cmd->argv[0], "echo") == 0)
-		return (builtin_echo(cmd));
-	if (ft_strcmp(cmd->argv[0], "exit") == 0)
-		return (builtin_exit(cmd));
-	if (ft_strcmp(cmd->argv[0], "env") == 0)
-		return (builtin_env(cmd));
-	if (ft_strcmp(cmd->argv[0], "pwd") == 0)
-		return (builtin_pwd(cmd));
-	if (ft_strcmp(cmd->argv[0], "export") == 0)
-		return (builtin_export(cmd));
-	if (ft_strcmp(cmd->argv[0], "unset") == 0)
-		return (builtin_unset(cmd));
-	return (0);
+	char **args;
+
+	args = node->args;
+	if (!args || !args[0])
+		return (g_exit_status = EXIT_FAILURE);
+	if (strcmp(args[0], "cd") == 0)
+		return (g_exit_status = builtin_cd(node, env));
+	if (strcmp(args[0], "export") == 0)
+		return (g_exit_status = builtin_export(node, env));
+	if (strcmp(args[0], "unset") == 0)
+		return (g_exit_status = builtin_unset(node, env));
+	if (strcmp(args[0], "env") == 0)
+		return (g_exit_status = builtin_env(node, env));
+	if (strcmp(args[0], "echo") == 0)
+		return (g_exit_status = builtin_echo(node));
+	if (strcmp(args[0], "exit") == 0)
+		builtin_exit(node);
+	return (g_exit_status = EXIT_FAILURE);
 }
