+++ src/builtin/builtin_echo.c
