+++ src/builtin/builtin_export.c
@@ -6,14 +6,14 @@
 /*   By: bewong <bewong@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/21 15:14:34 by bewong        #+#    #+#                 */
-/*   Updated: 2025/01/22 13:17:19 by bewong        ########   odam.nl         */
+/*   Updated: 2025/01/23 18:51:11 by bewong        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "executor.h"
 
 
-int	builtin_export(t_exec *cmd)
+int		builtin_export(t_ast_node *node, t_env **env)
 {
 	
 }
