+++ src/executor/error/error.c
@@ -1,33 +1,20 @@
 #include "builtin.h"
+#include "minishell.h"
+#include <stdio.h>
+#include <unistd.h>
 
-int error_msg(const char *msg, int include_errno)
+void error(const char *word, const char *msg)
 {
-	char *errno_msg;
-
-	write(STDERR_FILENO, "minishell: ", 11);
+	ft_putstr_fd(RED, STDERR_FILENO);
+	ft_putstr_fd(SHELL_ERROR, STDERR_FILENO);
 	if (msg)
-		write(STDERR_FILENO, msg, ft_strlen(msg));
-	if (include_errno)
 	{
-		errno_msg = strerror(errno);
-		if (errno_msg)
-		{
-			write(STDERR_FILENO, ": ", 2);
-			write(STDERR_FILENO, errno_msg, ft_strlen(errno_msg));
-		}
+		ft_putstr_fd(word, STDERR_FILENO);
+		ft_putstr_fd(": ", STDERR_FILENO);
+		ft_putendl_fd(msg, STDERR_FILENO);
 	}
-	write(STDERR_FILENO, "\n", 1);
-	return (1);
-}
-
-int error_msg_format(const char *prefix, const char *arg, const char *suffix)
-{
-	if (prefix)
-		printf("%s", prefix);
-	if (arg)
-		printf("%s", arg);
-	if (suffix)
-		printf("%s", suffix);
-	printf("\n");
-	return (1);
+	else
+		perror(word);
+	ft_putstr_fd(RESET, STDERR_FILENO);
+	update_exit_status(1, NULL);
 }
