+++ src/executor/utils.c/exit_update.c
@@ -23,6 +23,18 @@
 	4. Redirections (run_redir)
 		Update g_exit_status after applying redirections and running commands.
 */
+
+/*
+	Exit Code		Meaning 							Example
+	1				Catchall for general errors
+	2				Misuse of shell builtins			empty_function()
+	126				Command invoked cannot execute		/dev/null
+	127				"command not found"					illegal_command (typo)
+	128				Invalid argument to exit			kill -9 $PPID returns 137 (128 + 9)
+	130				Script terminated by Control-C		Ctl-C
+	255*			Exit status out of range			exit -1
+*/
+
 void	update_exit_status(int argc, char **args)
 {
 	if (argc == 0)
